bug: Fix subapp wildcard conflict issue
There is an issue when adding a subapp with a wildcard wherein sibling routes are not properly propagated to the new tree. This means that having these routes:

`/:id`
`/do-something`

makes only `/:id` propagate correctly. This PR fixes that by always propagating sibling routes regardless of the presence of a wildcard.