feat: Upgrade entire parse tree
Whew. What a lot of work this has been. Completely ripped out the parser and rebuilt it hopefully using a little more sanity in the code. This let me upgrade the way middleware works so it no longer checks a chain but rather builds a series of nested functions instead. In initial tests this has shown a roughly 8% improvement.